+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `--language` flag to `gen makefile`.
 - Generate main Makefile including `*.mk` files to allow custom Makefiles.
 - Pretty print errors.
+- Print devctl version in generated files headers.
 
 ## Changed